# perf: in-register lookup table & SIMD for 4bit PQ