… python/Cargo.lock
… rust/lance-index/src/vector/pq/storage.rs
… rust/lance-linalg/src/simd/f32.rs
… rust/lance/src/index/vector/ivf/v2.rs