4bit PQ is 3x faster than before:
```
16000,l2,PQ=96x4,DIM=1536
                        time:   [187.17 µs 187.95 µs 188.52 µs]
                        change: [-65.789% -65.641% -65.520%] (p = 0.00 < 0.10)
                        Performance has improved.

16000,cosine,PQ=96x4,DIM=1536
                        time:   [214.16 µs 214.52 µs 214.89 µs]
                        change: [-62.748% -62.594% -62.442%] (p = 0.00 < 0.10)
                        Performance has improved.

16000,dot,PQ=96x4,DIM=1536
                        time:   [190.12 µs 191.27 µs 192.22 µs]
                        change: [-65.496% -65.303% -65.086%] (p = 0.00 < 0.10)
                        Performance has improved.
```

post 8bit PQ results here for comparing, in short 4bit PQ is about 2x faster with the same index params:
```
compute_distances: 16000,l2,PQ=96,DIM=1536
                        time:   [405.11 µs 405.72 µs 406.92 µs]
                        change: [-0.2844% +0.1588% +0.6035%] (p = 0.50 > 0.10)
                        No change in performance detected.

compute_distances: 16000,cosine,PQ=96,DIM=1536
                        time:   [419.98 µs 421.05 µs 421.99 µs]
                        change: [-0.2540% +0.1098% +0.4928%] (p = 0.59 > 0.10)
                        No change in performance detected.

compute_distances: 16000,dot,PQ=96,DIM=1536
                        time:   [432.08 µs 433.63 µs 435.69 µs]
                        change: [-25.522% -25.243% -24.938%] (p = 0.00 < 0.10)
                        Performance has improved.
```